backend/voicecall-module/requirements.txt
fastapi
uvicorn[standard]
pipecat-ai[deepgram,cartesia,silero,websocket]
deepgram-sdk
cartesia
openai
anthropic
python-dotenv
pydantic
redis
httpx
websockets
